aegis/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aegis"
version = "0.1.0"
description = "Multi-agent RL for communication-deduction games inspired by Among Us"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "gymnasium>=0.29.0",
    "pettingzoo>=1.24.0",
    "supersuit>=3.9.0",
    "networkx>=3.0",
    "ray[rllib]>=2.9.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "wandb>=0.16.0",
    "pytest>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["aegis*"]

[tool.pytest.ini_options]
testpaths = ["aegis/tests"]
python_files = ["test_*.py"]

